Fix geoip parsing struct shape

+++ server/models/geoip_data.go
@@ -1,13 +1,13 @@
 package models
 
 type GeoIpData struct {
-	Range    [2]int `json:"range"`
-	Country  string `json:"country"`
-	Region   string `json:"region"`
-	Eu       string `json:"eu"`
-	Timezone string `json:"timezone"`
-	City     string `json:"city"`
-	Ll       [2]int `json:"ll"`
-	Metro    int    `json:"metro"`
-	Area     int    `json:"area"`
+	Range    [2]float64 `json:"range"`
+	Country  string     `json:"country"`
+	Region   string     `json:"region"`
+	Eu       string     `json:"eu"`
+	Timezone string     `json:"timezone"`
+	City     string     `json:"city"`
+	Ll       [2]float64 `json:"ll"`
+	Metro    int        `json:"metro"`
+	Area     int        `json:"area"`
 }
